Check verify email and send login data
