+++ src/config/mailConfig.js
@@ -56,9 +56,28 @@ const getTemplate = (name1Person, lastname1Person, token) => {
     `;
 }
 
+const getTemplatePassword = (name1Person, lastname1Person, emailPerson, passwordPerson) => {
+    return `
+      <head>
+          <link rel="stylesheet" href="./style.css">
+      </head>
+      
+      <div id="email___content">
+          <img src="" alt="">
+          <h2>Hola ${ name1Person + " "+ lastname1Person }</h2>
+          <p>Tu cuenta de correo ha sido verificada con exito. A continuación en contraras tu usuario y contraseña para acceder a la pagina de Programte School donde podras conocer e inscribirte a nuestras convocatorias.
+          </p>
+
+          <h6> Usuario: ${emailPerson} </h6>
+          <h6> Contraseña: ${passwordPerson} </h6>
+          
+      </div>
+    `;
+}
 
 
 module.exports={
     sendEmail,
     getTemplate,
+    getTemplatePassword,
 }
