--- src/config/mailPassword.js
@@ -1,64 +1,0 @@
-const nodemailer = require("nodemailer");
-require('dotenv').config( );
-
-//Here, the constant "mail" store the user and password of the SUPERADMIN o the email definite for programate school.
-
-//For this test, I used a google mail that is save in the .env, also the data in the constant 'transport'is the mailtrap.
-
-let transport = nodemailer.createTransport({
-    host: "smtp.gmail.com",
-    port: 465,
-    secure: true,
-    auth: {
-      user: process.env.userEmail,
-      pass: process.env.passEmail
-    },
-    tls: {
-        rejectUnauthorized: false
-    }
-});
-
-const sendPassword = async (emailPerson, subject, html ) => {
-    try {
-        await transport.sendMail({
-            from: ` ProgramateSchool <$( process.env.userEmail )>` ,// sender address
-            to: emailPerson, // list of receivers
-            subject,// Subject line
-            // text: "Holaaaa, vamos a verificar tu correo", // plain text body
-            html,
-        });
-
-    } catch (error) {
-        console.log('Algo no va bien con el envio de la contraseña', error);
-        // res
-        // .status (401)
-        // .send({status: 'Error con el email', data:sendEmail})     
-          
-    }     
-}
-
-const getTemplatePassword = (name1Person, lastname1Person, emailPerson, passwordPerson) => {
-    return `
-      <head>
-          <link rel="stylesheet" href="./style.css">
-      </head>
-      
-      <div id="email___content">
-          <img src="" alt="">
-          <h2>Hola ${ name1Person + " "+ lastname1Person }</h2>
-          <p>Tu cuenta de correo ha sido verificada con exito. A continuación en contraras tu usuario y contraseña para acceder a la pagina de Programte School donde podras conocer e inscribirte a nuestras convocatorias.
-          </p>
-
-          <h6> Usuario: ${emailPerson} </h6>
-          <h6> Contraseña: ${passwordPerson} </h6>
-          
-      </div>
-    `;
-}
-
-
-
-module.exports={
-    sendPassword,
-    getTemplatePassword,
-}
